Create camel-banana-puzzle.c
A person has 3000 bananas and a camel. The person wants to transport maximum number of bananas to a destination which

is 1000 KMs away, using only the camel as a mode of transportation. The camel cannot carry more than 1000 bananas at a time

and eats a banana every km it travels. What is the maximum number of bananas that can be transferred to the destination using

only camel